refactor: move query and wait logic to separate module
This prepares the way for using the `query_and_wait` method built-in to the
client library when available.
